web/cdk/requirements.txt
# AWS CDK Dependencies
aws-cdk-lib>=2.170.0,<3.0.0
constructs>=10.0.0,<11.0.0

# Python utilities
python-dotenv>=1.0.0,<2.0.0
